Rename fixtures file
